requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
Cython==0.29.14
decorator==4.4.2
filelock==3.0.12
future==0.18.2
gensim==3.8.3
huggingface-hub==0.0.12
hyperopt==0.2.5
idna==2.10
importlib-metadata==4.6.0
joblib==1.0.1
jsonlines==2.0.0
munkres==1.1.4
networkx==2.5.1
nltk==3.6.2
numpy==1.21.0
Pillow==8.2.0
packaging==20.9
protobuf==3.17.3
psutil==5.8.0
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.1.0
stanza==1.2.1
Theano==1.0.5
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.1
transformers==4.8.2
truecase==0.0.14
typing-extensions==3.10.0.0
urllib3==1.26.6
wincertstore==0.2
xdg==5.1.0
zipp==3.4.1
